Refactor NewRequestHeaderCallback out of GatewayBridge
