+++ dd-java-agent/appsec/src/main/java/com/datadog/appsec/gateway/GatewayBridge.java
@@ -13,7 +13,6 @@ import com.datadog.appsec.event.data.KnownAddresses;
 import com.datadog.appsec.event.data.MapDataBundle;
 import com.datadog.appsec.event.data.ObjectIntrospection;
 import com.datadog.appsec.event.data.SingletonDataBundle;
-import datadog.trace.api.function.TriConsumer;
 import datadog.trace.api.function.TriFunction;
 import datadog.trace.api.gateway.Events;
 import datadog.trace.api.gateway.Flow;
@@ -306,24 +305,6 @@ public class GatewayBridge {
 
   public void stop() {
     subscriptionService.reset();
-  }
-
-  private static class NewRequestHeaderCallback
-      implements TriConsumer<RequestContext, String, String> {
-    @Override
-    public void accept(RequestContext ctx_, String name, String value) {
-      AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
-      if (ctx == null) {
-        return;
-      }
-
-      if (name.equalsIgnoreCase("cookie")) {
-        Map<String, List<String>> cookies = CookieCutter.parseCookieHeader(value);
-        ctx.addCookies(cookies);
-      } else {
-        ctx.addRequestHeader(name, value);
-      }
-    }
   }
 
   private class MethodAndRawURICallback
